% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f2011}
\alias{f2011}
\title{\code{f2011}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 58478 rows and 85 columns.}
\source{
\url{https://meps.ahrq.gov/mepsweb/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-145}
}
\usage{
f2011
}
\description{
MEPS 2011 Person Round Plan Public Use File
}
\keyword{datasets}
