% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f2012}
\alias{f2012}
\title{\code{f2012}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 63361 rows and 85 columns.}
\source{
\url{https://meps.ahrq.gov/mepsweb/data_stats/download_data_files_detail.jsp?cboPufNumber=HC-153}
}
\usage{
f2012
}
\description{
MEPS 2012 Person Round Plan Public Use File
}
\keyword{datasets}
